requirements.txt
fastapi
uvicorn
pdfminer.six
python-multipart
requests